Fix SQL number precision

+++ components/Main.tsx
@@ -95,7 +95,7 @@ export default function Main() {
             `CAST("${field}" AS ${
               getTypeForColumn(queryFields, files[0].name, field) === "Utf8"
                 ? "VARCHAR"
-                : "DECIMAL"
+                : "DOUBLE"
             }) AS "${field}"`
         )
         .join(", ");
@@ -105,7 +105,7 @@ export default function Main() {
             `CAST("${field}" AS ${
               getTypeForColumn(queryFields, files[0].name, field) === "Utf8"
                 ? "VARCHAR"
-                : "DECIMAL"
+                : "DOUBLE"
             })`
         )
         .join(", ");
@@ -117,7 +117,7 @@ export default function Main() {
         getTypeForColumn(queryFields, files[0].name, aggregation.name) ===
         "Utf8"
           ? "VARCHAR"
-          : "DECIMAL"
+          : "DOUBLE"
       })) AS "${aggregation.name}"
           FROM '${files[0].name}' 
           ${
@@ -152,7 +152,7 @@ export default function Main() {
           }" AS ${
             getTypeForColumn(queryFields, field.table, field.name) === "Utf8"
               ? "VARCHAR"
-              : "DECIMAL"
+              : "DOUBLE"
           }) AS "${field.name}"`
       );
 
@@ -164,7 +164,7 @@ export default function Main() {
             )}."${field.name}" AS ${
               getTypeForColumn(queryFields, field.table, field.name) === "Utf8"
                 ? "VARCHAR"
-                : "DECIMAL"
+                : "DOUBLE"
             })`
         )
         .join(", ");
@@ -183,7 +183,7 @@ export default function Main() {
             relationship.foreign_key
           ) !== "Utf8";
         const castType =
-          isPrimaryFloat || isForeignFloat ? "DECIMAL" : "VARCHAR";
+          isPrimaryFloat || isForeignFloat ? "DOUBLE" : "VARCHAR";
 
         return `CAST(TABLE${files.findIndex(
           (file) => file.name === relationship.primary_table
@@ -205,7 +205,7 @@ export default function Main() {
           aggregation.name
         ) === "Utf8"
           ? "VARCHAR"
-          : "DECIMAL"
+          : "DOUBLE"
       })) AS "${aggregation.name}"
           FROM '${files[0].name}' AS TABLE0
           JOIN ${files
